data changes

+++ src/SFA.DAS.Payments.MatchedLearner.Data/Entities/PeriodModel.cs
@@ -5,12 +5,12 @@
         public long Id { get; set; }
         public long PriceEpisodeId { get; set; }
         public bool IsPayable { get; set; }
-        public short TransactionType { get; set; }
+        public byte TransactionType { get; set; }
         public byte Period { get; set; }
         public decimal Amount { get; set; }
         public long? ApprenticeshipId { get; set; }
         public long? AccountId { get; set; }
-        public int? ApprenticeshipEmployerType { get; set; }
+        public byte? ApprenticeshipEmployerType { get; set; }
         public long? TransferSenderAccountId { get; set; }
         public bool FailedDataLock1 { get; set; }
         public bool FailedDataLock2 { get; set; }
